moved devops logging utillity to build pipeline namespace (#885)
disable devops logging when ignore compiler errors is set in build args
