+++ XRTK-Core/Packages/com.xrtk.core/Editor/BuildPipeline/BuildInfo.cs
@@ -154,6 +154,9 @@ namespace XRTK.Editor.BuildPipeline
             {
                 switch (arguments[i])
                 {
+                    case "-ignoreCompilerErrors":
+                        DevOpsLoggingUtility.LoggingEnabled = false;
+                        break;
                     case "-autoIncrement":
                         AutoIncrement = true;
                         break;
