+++ XRTK-Core/Packages/com.xrtk.core/Tests/Core/TestFixture_01_MixedRealityToolkitTests.cs
@@ -6,6 +6,7 @@ using UnityEditor.SceneManagement;
 using UnityEngine;
 using UnityEngine.TestTools;
 using XRTK.Definitions;
+using XRTK.Editor.BuildPipeline;
 using XRTK.Editor.Utilities;
 using XRTK.Interfaces;
 using XRTK.Services;
