+++ XRTK-Core/Packages/com.xrtk.core/Tests/XRTK.Tests.asmdef
@@ -6,7 +6,8 @@
         "GUID:e67d30660ec243e4836aac191d3f36fb",
         "GUID:60a1dda7b8376464f89e6b89eb538044",
         "GUID:27619889b8ba8c24980f49ee34dbb44a",
-        "GUID:0acc523941302664db1f4e527237feb3"
+        "GUID:0acc523941302664db1f4e527237feb3",
+        "GUID:258402e6b12e10c43925784eabc239e5"
     ],
     "includePlatforms": [
         "Editor"
